+++ src/common/service.h
@@ -9,7 +9,7 @@ class IService {
 public:
     virtual ~IService() = default;
 
-    virtual int cur_interval() = 0;
+    virtual double cur_interval() = 0;
     virtual bool has_more() = 0;
     virtual bool is_done() = 0;
     virtual bool reserve() = 0;
@@ -20,10 +20,10 @@ public:
 
 class AbstractService: public IService {
 public:
-    AbstractService(int round_cnt, int interval, util::Noises noises);
+    AbstractService(int round_cnt, double interval, util::Noises noises);
     ~AbstractService() override = default;
 
-    int cur_interval() override;
+    double cur_interval() override;
     bool has_more() override;
     bool is_done() override = 0;
     bool reserve() override;
@@ -33,7 +33,7 @@ public:
 
 protected:
     int round_cnt;
-    int interval;
+    double interval;
     util::Noises noises;
 
     std::atomic<int> cur_idx_reserved;
